+++ test/test_ethernet/test_main.cpp
@@ -255,9 +255,9 @@ static bool waitForLocalIP() {
   TEST_MESSAGE(format("     Netmask: %u.%u.%u.%u", netmask[0], netmask[1], netmask[2], netmask[3]).data());
   TEST_MESSAGE(format("     Gateway: %u.%u.%u.%u", gateway[0], gateway[1], gateway[2], gateway[3]).data());
   TEST_MESSAGE(format("     DNS:     %u.%u.%u.%u", dns[0], dns[1], dns[2], dns[3]).data());
-  for (int i = 0; i < DNSClient::maxServers(); ++i) {
+  for (size_t i = 0; i < DNSClient::maxServers(); ++i) {
     dns = DNSClient::getServer(i);
-    TEST_MESSAGE(format("DNS Server %d: %u.%u.%u.%u", i, dns[0], dns[1], dns[2], dns[3]).data());
+    TEST_MESSAGE(format("DNS Server %zu: %u.%u.%u.%u", i, dns[0], dns[1], dns[2], dns[3]).data());
   }
 
   return true;